refactor: split xtask hardware diagnostics
